Allow for passing in of port

+++ src/rpc/client/ts/runTests.ts
@@ -1,9 +1,10 @@
-const { execSync, spawn } = require('child_process')
+const { execFileSync, spawn } = require('child_process')
 const { exit } = require('process')
 const glob = require('glob')
 const fs = require('fs')
 const os = require('os')
 const unzip = require('unzip-stream')
+const port = process.env.PORT || '9000'
 
 // Extract server
 async function extractServer(filePath) {
@@ -24,14 +25,14 @@ async function extractServer(filePath) {
 // Run Scala gRPC server
 async function startServer(filePath) {
   if (!os.platform().toLowerCase().startsWith('win')) {
-    execSync(`chmod +x ${filePath}/bin/omega-edit-grpc-server`)
+    execFileSync('chmod', ['+x', `${filePath}/bin/omega-edit-grpc-server`])
   }
 
   let scriptName = os.platform().toLowerCase().startsWith('win')
     ? `./${filePath}/bin/omega-edit-grpc-server.bat`
     : `./${filePath}/bin/omega-edit-grpc-server`
 
-  const server_process = spawn(scriptName, [], {
+  const server_process = spawn(scriptName, [`--port=${port}`], {
     stdio: 'ignore',
     detached: true,
   })
@@ -39,23 +40,40 @@ async function startServer(filePath) {
   fs.writeFileSync('server_pid', server_process.pid.toString())
 }
 
+// Method to getFilePath based on the name of the server package
+async function getFilePath() {
+  var serverFilePaths = await glob.sync('omega-edit-grpc-server-*', {
+    cwd: '.',
+  })
+
+  var serverFilePath = ''
+
+  for (var i = 0; i < serverFilePaths.length; i++) {
+    if (serverFilePaths[i].includes('.zip')) {
+      serverFilePath = serverFilePaths[i].replace('.zip', '')
+      break
+    }
+  }
+
+  return serverFilePath
+}
+
 // Stop Scala gRPC server
 async function stopServer() {
+  var serverFilePath = await getFilePath()
+  if (serverFilePath === '') exit(1)
+
   if (fs.existsSync('server_pid')) {
     process.kill(fs.readFileSync('server_pid').toString())
   }
 
-  fs.rmdirSync('omega-edit-grpc-server-0.9.20', { recursive: true })
+  fs.rmdirSync(serverFilePath, { recursive: true })
   fs.rmSync('server_pid')
 }
 
 // Run server by first extracting server then starting it
 async function runScalaServer() {
-  var serverFilePath = await glob
-    .sync('omega-edit-grpc-server-*', { cwd: '.' })
-    .toString()
-    .replace('.zip', '')
-
+  var serverFilePath = await getFilePath()
   if (serverFilePath === '') exit(1)
 
   await extractServer(serverFilePath)
@@ -66,4 +84,10 @@ async function runScalaServer() {
 module.exports = {
   stopScalaServer: stopServer,
   runScalaServer: runScalaServer,
+}
+
+if (process.argv.includes('runScalaServer')) {
+  runScalaServer()
+} else if (process.argv.includes('stopScalaServer')) {
+  stopServer()
 }
